后台管理  商品介绍图片预览 小程序  患者登录 通过openId,unionId登录，返回用户信息

+++ jingjingshop/src/main/java/com/pinyougou/mapper/TbPatientMapper.java
@@ -22,7 +22,7 @@ public interface TbPatientMapper {
 
     List<TbPatient> selectByExample(TbPatientExample example);
 
-    TbPatient selectByPrimaryKey(Integer pid);
+    TbPatient selectByPrimaryKey(Long pid);
 
     int updateByExampleSelective(@Param("record") TbPatient record, @Param("example") TbPatientExample example);
